+++ src/components/member.vue
@@ -14,7 +14,7 @@
     {{ memb.email }}
   </span>
 
-  <span class="panel-icon">
+  <span v-if="ok" class="panel-icon">
     <a @click="suprMember">
       <i class="icon-trash"> </i>
     </a>
@@ -34,16 +34,22 @@ export default {
   data () {
     return {
       result: '',
+      ok: false,
     }
   },
   mounted () {
     this.result = 'https://www.gravatar.com/avatar/'+md5(this.memb.email)+'?d=https://api.adorable.io/avatars/285/abott@adorable.png';
+    if(this.memb._id != this.$store.state.member._id)
+      this.ok = true;
+
   },
   methods : {
     suprMember() {
-
+      if(confirm("Voulez vous supprimer le membre ?"))
+      {
         window.axios.delete('members/'+this.memb._id);
         window.bus.$emit('updateMember');
+      }
     }
   }
